Fixing lot of win64 type conversion warnings (in progress)
Former-commit-id: 4ee5fa10fd43bc451242bba96b358f70536bdc46

+++ framework/sofa/core/objectmodel/BaseLink.cpp
@@ -66,7 +66,7 @@ void BaseLink::printValue( std::ostream& o ) const
     bool first = true;
 	for (std::size_t i = 0; i<size; ++i)
     {
-        std::string path = getLinkedPath(i);
+        std::string path = getLinkedPath((unsigned int)i);
         if (path.empty()) continue;
         if (first) first = false;
         else o << ' ';
